// src/app/models/sign-up-model.ts
export class SignUpModel {
  Id: number;
  LoginEmail: string;
  FirstName: string;
  LastName: string;
  Phone: string;
  Prefix: string;
  UserName: string;
  //CreatedById: number;
  //UpdatedById: number;
  //CreatedDateTime: Date;
  //UpdatedDateTime: Date;
  CountryName: string;
  JobTitle: string;
  CompanyName: string;
  PhoneNumber: string;
  Country: string;
}